requirements.txt
openpyxl>=3.1.2
python-docx>=1.1.0
pypdf>=4.2.0
